+++ src/plugins/contrib-wip/cbMemoryView/MemoryPanel.h
@@ -46,7 +46,7 @@ class MemoryPanel: public wxPanel
         //*)
 
         wxString m_sAddress;
-        uint64_t m_llSize;
+        long long unsigned int m_llSize;
         std::shared_ptr<cbWatch> m_watch;
 
         DECLARE_EVENT_TABLE()
